requirements.txt
PyMuPDF
sentence-transformers
faiss-cpu
llama-index
